Rename : mypage도메인 dto 디렉토리 구조 변경
