+++ src/main/java/noodlezip/badge/service/BadgeServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import noodlezip.badge.exception.BadgeErrorStatus;
 import noodlezip.badge.repository.BadgeRepository;
 import noodlezip.common.exception.CustomException;
-import noodlezip.mypage.dto.response.LevelBadgeDetailResponse;
+import noodlezip.mypage.dto.response.badge.LevelBadgeDetailResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
