+++ src/main/java/noodlezip/badge/service/UserBadgeService.java
@@ -1,8 +1,8 @@
 package noodlezip.badge.service;
 
-import noodlezip.mypage.dto.UserNoOptionBadgeDto;
-import noodlezip.mypage.dto.UserOptionBadgeDto;
-import noodlezip.mypage.dto.response.BadgeGroupResponse;
+import noodlezip.badge.dto.UserNoOptionBadgeDto;
+import noodlezip.badge.dto.UserOptionBadgeDto;
+import noodlezip.mypage.dto.response.badge.BadgeGroupResponse;
 
 import java.util.List;
