+++ src/main/java/noodlezip/badge/service/UserBadgeServiceImpl.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import noodlezip.badge.constants.LevelBadgeCategoryType;
 import noodlezip.badge.repository.BadgeGroupRepository;
 import noodlezip.badge.repository.UserBadgeRepository;
-import noodlezip.mypage.dto.UserNoOptionBadgeDto;
-import noodlezip.mypage.dto.UserOptionBadgeDto;
-import noodlezip.mypage.dto.response.BadgeGroupResponse;
+import noodlezip.badge.dto.UserNoOptionBadgeDto;
+import noodlezip.badge.dto.UserOptionBadgeDto;
+import noodlezip.mypage.dto.response.badge.BadgeGroupResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
