--- src/main/java/noodlezip/mypage/dto/response/SavedStoreCategoryFilterResponse.java
@@ -1,12 +1,0 @@
-package noodlezip.mypage.dto.response;
-
-import noodlezip.savedstore.dto.response.SavedStoreCategoryResponse;
-
-import java.util.List;
-
-public class SavedStoreCategoryFilterResponse {
-
-    private List<SavedStoreCategoryResponse> savedStoreCategoryList;
-
-}
-
