--- src/main/java/noodlezip/mypage/dto/response/SavedStoreListResponse.java
@@ -1,13 +1,0 @@
-package noodlezip.mypage.dto.response;
-
-import java.util.List;
-
-public class SavedStoreListResponse {
-    /// 체크표시한 가게 pk도 넘겨야함
-    ///  페이지네이션 정보
-    ///  가게 리스트(위치가 포함된) -> 이거는 따로
-    
-    private List<SavedStoreResponse> savedStoreList;
-    /// 페이지네이션 정보
-    
-}
