--- src/main/java/noodlezip/mypage/dto/response/SavedStoreResponse.java
@@ -1,13 +1,0 @@
-package noodlezip.mypage.dto.response;
-
-public class SavedStoreResponse {
-
-    private Long savedStoreId;
-    private Long saveStoreCategoryId;
-    private Long storeId;
-    private String storeName;
-    private String address;
-    private String storeMainImageUrl;
-    private String memo;
-
-}
