+++ src/main/java/noodlezip/mypage/dto/response/badge/BadgeDetailResponse.java
@@ -1,4 +1,4 @@
-package noodlezip.mypage.dto.response;
+package noodlezip.mypage.dto.response.badge;
 
 import lombok.*;
