+++ src/main/java/noodlezip/mypage/service/MyBadgeServiceImpl.java
@@ -11,9 +11,9 @@ import noodlezip.badge.service.BadgeService;
 import noodlezip.badge.service.UserBadgeService;
 import noodlezip.common.exception.CustomException;
 import noodlezip.common.status.ErrorStatus;
-import noodlezip.mypage.dto.UserNoOptionBadgeDto;
-import noodlezip.mypage.dto.UserOptionBadgeDto;
-import noodlezip.mypage.dto.response.*;
+import noodlezip.badge.dto.UserNoOptionBadgeDto;
+import noodlezip.badge.dto.UserOptionBadgeDto;
+import noodlezip.mypage.dto.response.badge.*;
 import noodlezip.user.service.UserService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
