+++ src/test/java/noodlezip/badge/service/BadgeServiceImplTest.java
@@ -3,7 +3,7 @@ package noodlezip.badge.service;
 import noodlezip.badge.exception.BadgeErrorStatus;
 import noodlezip.badge.repository.BadgeRepository;
 import noodlezip.common.exception.CustomException;
-import noodlezip.mypage.dto.response.LevelBadgeDetailResponse;
+import noodlezip.mypage.dto.response.badge.LevelBadgeDetailResponse;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
