+++ src/test/java/noodlezip/badge/service/UserBadgeServiceImplTest.java
@@ -3,9 +3,9 @@ package noodlezip.badge.service;
 import noodlezip.badge.constants.LevelBadgeCategoryType;
 import noodlezip.badge.repository.BadgeGroupRepository;
 import noodlezip.badge.repository.UserBadgeRepository;
-import noodlezip.mypage.dto.UserNoOptionBadgeDto;
-import noodlezip.mypage.dto.UserOptionBadgeDto;
-import noodlezip.mypage.dto.response.BadgeGroupResponse;
+import noodlezip.badge.dto.UserNoOptionBadgeDto;
+import noodlezip.badge.dto.UserOptionBadgeDto;
+import noodlezip.mypage.dto.response.badge.BadgeGroupResponse;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
